Significantly overhauled the structure of array_t to make it more flexible when using programmatically.
Added a corresponding array_span, which is a non-owning span into an array_t, which can index into array_t.

+++ grad.h
@@ -22,8 +22,8 @@ class grad_class {
     const jax::expression& input_expr;
     jax::expression output_expr;
 
-    std::unordered_map<u32, bool> active_adjoint_map;
-    std::unordered_map<u32, jax::value> variable_adjoint_map;
+    std::unordered_map<size_t, bool> active_adjoint_map;
+    std::unordered_map<size_t, jax::value> variable_adjoint_map;
 };
 
 jax::expression grad(const jax::expression& expr);
